pyproject.toml
[project]
name = "PMD_extraction_and_analysis"
version = "0.1.0"
description = "A fully automtic pipeline for processing perfusion MRI scans. This pipeline extracts explainable image feature; no deeplearning or strong PC required. Perfect for when DL-based feature extraction is not possible due to small datasets or hardware limitations. These explainable features can be used to perform statistical analysis or traditional machine learning (also provided in this package)."
dependencies = [
    "antspyx==0.5.4",
    "asgiref==3.8.1",
    "certifi==2025.4.26",
    "cffi==1.17.1",
    "charset-normalizer==3.4.2",
    "contourpy==1.3.2",
    "cycler==0.12.1",
    "diptest==0.9.0",
    "Django==5.2.1",
    "et_xmlfile==2.0.0",
    "fonttools==4.58.0",
    "fsspec==2025.3.2",
    "gevent==25.4.2",
    "greenlet==3.2.2",
    "idna==3.10",
    "importlib_resources==6.5.2",
    "joblib==1.5.0",
    "kiwisolver==1.4.8",
    "matplotlib==3.10.3",
    "nibabel==5.3.2",
    "numpy==2.0.1",
    "openpyxl==3.1.5",
    "packaging==25.0",
    "pandas==2.2.3",
    "patsy==1.0.1",
    "pillow==11.2.1",
    "psutil==7.0.0",
    "pycparser==2.22",
    "pyparsing==3.2.3",
    "python-dateutil==2.9.0.post0",
    "pytz==2025.2",
    "PyYAML==6.0.2",
    "requests==2.32.3",
    "scikit-learn==1.6.1",
    "scipy==1.15.3",
    "seaborn==0.13.2",
    "six==1.17.0",
    "sqlparse==0.5.3",
    "statsmodels==0.14.4",
    "threadpoolctl==3.6.0",
    "typing_extensions==4.13.2",
    "tzdata==2025.2",
    "urllib3==2.4.0",
    "webcolors==24.11.1",
    "xarray==2025.4.0",
    "zope.event==5.0",
    "zope.interface==7.2"
]
